requirements.txt
requests
openai
faiss-cpu
python-dotenv
numpy
PyPDF2
streamlit